Downgrade to SimpleLink CC13xx and CC26xx SDK 7.10.00.98
The examples repository was missing a few things in the past, and this
commit was the closest (the first one) that relates to what is shipped
with SDK 7.10.00.98. So lets match the SDK with that.

Future commits/tags are more up-to-date so we can be close to what is
needed then.

Note, that it is obviously needed to update the submodule's URL to point
to our own fork so that we can add patches.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>